% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contact_table}
\alias{contact_table}
\title{DATASET_TITLE}
\format{A data frame with 16859 rows and 5 variables:
\describe{
  \item{\code{employee_num}}{integer COLUMN_DESCRIPTION}
  \item{\code{contact_type}}{character COLUMN_DESCRIPTION}
  \item{\code{contact_sub_type}}{character COLUMN_DESCRIPTION}
  \item{\code{contact}}{character COLUMN_DESCRIPTION}
  \item{\code{contact_end_date}}{double COLUMN_DESCRIPTION} 
}}
\source{
\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}
}
\usage{
contact_table

contact_table

contact_table

contact_table
}
\description{
DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION
}
\keyword{datasets}
