% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deskhistory_table}
\alias{deskhistory_table}
\title{Desk History}
\format{A data frame with 4724 rows and 6 variables:
\describe{
  \item{\code{employee_num}}{integer COLUMN_DESCRIPTION}
  \item{\code{desk_id}}{integer COLUMN_DESCRIPTION}
  \item{\code{desk_id_start_date}}{double COLUMN_DESCRIPTION}
  \item{\code{desk_id_end_date}}{double COLUMN_DESCRIPTION}
  \item{\code{termination_flag}}{integer COLUMN_DESCRIPTION}
  \item{\code{promotion_flag}}{integer COLUMN_DESCRIPTION}
}}
\source{
\url{https://github.com/harryahlas/hrsample/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}
}
\usage{
deskhistory_table

deskhistory_table

deskhistory_table

deskhistory_table

deskhistory_table

deskhistory_table

deskhistory_table
}
\description{
History of employees in each position at fictional company.

DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION
}
\keyword{datasets}
