% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deskjob_table}
\alias{deskjob_table}
\title{Desk Job Table}
\format{A data frame with 1212 rows and 2 variables:
\describe{
  \item{\code{desk_id}}{integer COLUMN_DESCRIPTION}
  \item{\code{job_name}}{character COLUMN_DESCRIPTION}
}}
\source{
\url{https://github.com/harryahlas/hrsample/}
}
\usage{
deskjob_table
}
\description{
Assigns a job title to each desk in fictional company.
}
\keyword{datasets}
