% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{education_table}
\alias{education_table}
\title{DATASET_TITLE}
\format{A data frame with 3917 rows and 3 variables:
\describe{
  \item{\code{employee_num}}{integer COLUMN_DESCRIPTION}
  \item{\code{degree}}{character COLUMN_DESCRIPTION}
  \item{\code{school_name}}{character COLUMN_DESCRIPTION} 
}}
\source{
\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}
}
\usage{
education_table

education_table

education_table

education_table
}
\description{
DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION
}
\keyword{datasets}
