% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrsampleCreateSQLServer.R
\name{hrsampleCreateSQLServer}
\alias{hrsampleCreateSQLServer}
\title{Load hrsample to SQL Server}
\usage{
hrsampleCreateSQLServer(driver = "ODBC Driver 13 for SQL Server",
  dbname = "master", server = NULL, user = NULL, password = NULL,
  port = 1433)
}
\arguments{
\item{driver}{Driver that will be used to connect to SQL Server.  Might be something like 'ODBC Driver 13 for SQL Server'.}

\item{dbname}{The name of the database the tables will be stored to.  Must be a preexisting database. Ideal name would be "hrsample". This function does not create a new database.}

\item{server}{Name of the SQL server. If working with a remote database, may be something like "johnny.heliohost.org".  If working with a database on your local machine it might be "localhost\\\\SQLEXPRESS".}

\item{user}{Username to log into database}

\item{password}{Password for user}

\item{port}{port to be used}
}
\description{
This function loads the hrsample tables to a Microsoft SQL Server database that the user has access to. It utilizes the odbc package .  The parameters are based on the odbc parameters.
}
\examples{
hrsampleCreateSQLServer()
}
\keyword{hrsample}
