% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{performancereview_table}
\alias{performancereview_table}
\title{Performance Review Table}
\format{A data frame with 11661 rows and 3 variables:
\describe{
  \item{\code{employee_num}}{integer COLUMN_DESCRIPTION}
  \item{\code{year}}{integer COLUMN_DESCRIPTION}
  \item{\code{perf_review_score}}{integer COLUMN_DESCRIPTION}
}}
\source{
\url{https://github.com/harryahlas/hrsample/}
}
\usage{
performancereview_table
}
\description{
Annual performance reviews at fictional company.
}
\keyword{datasets}
