% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recruiting_table}
\alias{recruiting_table}
\title{DATASET_TITLE}
\format{A data frame with 4509 rows and 2 variables:
\describe{
  \item{\code{employee_num}}{integer COLUMN_DESCRIPTION}
  \item{\code{recruiting_source}}{character COLUMN_DESCRIPTION} 
}}
\source{
\url{http://somewhere.important.com/}
}
\usage{
recruiting_table
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
