% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rollup_view}
\alias{rollup_view}
\title{DATASET_TITLE}
\format{A data frame with 1233 rows and 11 variables:
\describe{
  \item{\code{DESK_ID_LEVEL_0}}{integer COLUMN_DESCRIPTION}
  \item{\code{HRC_LEVEL_0}}{character COLUMN_DESCRIPTION}
  \item{\code{DESK_ID_LEVEL_1}}{integer COLUMN_DESCRIPTION}
  \item{\code{HRC_LEVEL_1}}{character COLUMN_DESCRIPTION}
  \item{\code{DESK_ID_LEVEL_2}}{integer COLUMN_DESCRIPTION}
  \item{\code{HRC_LEVEL_2}}{character COLUMN_DESCRIPTION}
  \item{\code{DESK_ID_LEVEL_3}}{integer COLUMN_DESCRIPTION}
  \item{\code{HRC_LEVEL_3}}{character COLUMN_DESCRIPTION}
  \item{\code{DESK_ID_LEVEL_4}}{integer COLUMN_DESCRIPTION}
  \item{\code{HRC_LEVEL_4}}{character COLUMN_DESCRIPTION}
  \item{\code{depth}}{double COLUMN_DESCRIPTION} 
}}
\source{
\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}
}
\usage{
rollup_view

rollup_view

rollup_view

rollup_view

rollup_view

rollup_view

rollup_view
}
\description{
DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION
}
\keyword{datasets}
