% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{salaryhistory_table}
\alias{salaryhistory_table}
\title{Salary History Table}
\format{A data frame with 13350 rows and 5 variables:
\describe{
  \item{\code{employee_num}}{integer COLUMN_DESCRIPTION}
  \item{\code{salary_effective_date}}{double COLUMN_DESCRIPTION}
  \item{\code{salary}}{double COLUMN_DESCRIPTION}
  \item{\code{salary_increase}}{double COLUMN_DESCRIPTION}
  \item{\code{starting_salary_flag}}{character COLUMN_DESCRIPTION}
}}
\source{
\url{https://github.com/harryahlas/hrsample/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}
}
\usage{
salaryhistory_table

salaryhistory_table

salaryhistory_table

salaryhistory_table

salaryhistory_table

salaryhistory_table
}
\description{
Salary history at fictitional company.

DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION
}
\keyword{datasets}
