% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{skills_table}
\alias{skills_table}
\title{DATASET_TITLE}
\format{A data frame with 5015 rows and 3 variables:
\describe{
  \item{\code{employee_num}}{integer COLUMN_DESCRIPTION}
  \item{\code{skill_name}}{character COLUMN_DESCRIPTION}
  \item{\code{skill_type}}{character COLUMN_DESCRIPTION} 
}}
\source{
\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}
}
\usage{
skills_table

skills_table

skills_table

skills_table

skills_table
}
\description{
DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION
}
\keyword{datasets}
