% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{employeeinfo_table}
\alias{employeeinfo_table}
\title{Employee information}
\format{A data frame with 45000 rows and 6 variables:
\describe{
  \item{\code{employee_num}}{integer COLUMN_DESCRIPTION}
  \item{\code{first_name}}{character COLUMN_DESCRIPTION}
  \item{\code{last_name}}{character COLUMN_DESCRIPTION}
  \item{\code{hire_date}}{double COLUMN_DESCRIPTION}
  \item{\code{city}}{character COLUMN_DESCRIPTION}
  \item{\code{state}}{character COLUMN_DESCRIPTION}
}}
\source{
\url{https://github.com/harryahlas/hrsample/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}
}
\usage{
employeeinfo_table

employeeinfo_table

employeeinfo_table

employeeinfo_table

employeeinfo_table

employeeinfo_table

employeeinfo_table
}
\description{
Fictional information about current, former, and potential company employees

DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION
}
\keyword{datasets}
