% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hierarchy_table}
\alias{hierarchy_table}
\title{Hierarchy Table}
\format{A data frame with 1212 rows and 3 variables:
\describe{
  \item{\code{desk_id}}{integer COLUMN_DESCRIPTION}
  \item{\code{org}}{character COLUMN_DESCRIPTION}
  \item{\code{parent_id}}{integer COLUMN_DESCRIPTION}
}}
\source{
\url{https://github.com/harryahlas/hrsample/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}

\url{http://somewhere.important.com/}
}
\usage{
hierarchy_table

hierarchy_table

hierarchy_table

hierarchy_table

hierarchy_table

hierarchy_table

hierarchy_table
}
\description{
Shows which position each position reports to in fictional company.

DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION

DATASET_DESCRIPTION
}
\keyword{datasets}
